Move DictionaryCache
