+++ synapse/util/dictionary_cache.py
